+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Build 1473 - 2026-08-24
+
+- automated build
+
 ## Build 1472 - 2026-08-24
 
 - automated build